Change install paths to place .kext in /Library/Extensions instead of /System/Library/Extensions
Although previously, unsigned kexts needed to be installed in /System/Library/Extensions
to even have a chance of installing (by allowing unsigned kexts to be loaded)
now OS X (As of El Capitan) does not allow loading these.
Now the kexts need to be signed, and placed in /Library/Extensions for them to load.
(Unless you grant yourself rootless access and such)